Combined string.h for user and kernel lands
