+++ src/userio.h
@@ -1,0 +1,13 @@
+/*
+Copyright (C) 2015 The University of Notre Dame
+This software is distributed under the GNU General Public License.
+See the file LICENSE for details.
+*/
+
+#ifndef USTRING_H
+#define USTRING_H
+
+void printf_putchar( char c );
+void printf_putstring( char *s );
+
+#endif
